--- Sequence_Calculator.depend
@@ -1,8 +1,0 @@
-# depslib dependency file v1.0
-1633595949 source:d:\documents\codes\codeblocks\c++ projects\calculator\sequence_calculator\main.cpp
-	<iostream>
-	<vector>
-	<cmath>
-
-1633596338 source:d:\documents\codes\codeblocks\c++ projects\calculator\sequence_calculator\multi_stage.c
-
