+++ geometric.cpp
@@ -1,0 +1,90 @@
+#include "geometric.hpp"
+
+bool Geometric::is_exponential(const vector<double>& seq)
+{
+    for(int i = 1; i<seq.size(); ++i){
+        rate.push_back(seq[i]/seq[i-1]);
+    }
+
+    for(int i = 0; i<rate.size(); ++i){
+        if(rate[0] == rate[i]){
+            constant = true;
+        } else {
+            constant = false;
+            break;
+        }
+    }
+    return constant;
+
+    /*if(constant){
+        return constant;
+    }
+
+    for(int i = 1; i<rate.size(); ++i){
+        rate2.push_back(rate[i]/rate[i-1]);
+    }
+
+    for(int i = 0; i<rate2.size(); ++i){
+        if(rate2[0] == rate2[i]){
+            Multi_stage_type = true;
+        } else {
+            Multi_stage_type = false;
+            return false;
+        }
+    }
+    return true;*/
+}
+
+double Geometric::geometric(vector<double>& seq, const int& n)
+{
+    double a1 = seq[0];
+    double r = seq[1]/seq[0];
+    double ans;   // an = a1*ar^n-1
+
+    if(constant){
+        ans = a1*pow(r,(n-1));
+        if(n > seq.size()){
+            for(int i = seq.size()-1; i != n-1; ++i){
+                seq.push_back(seq[i]*r);
+            }
+        }
+    } /*else {
+        r = (seq[seq.size()-1] / seq[seq.size()-2]) * rate2[0];
+        if(n > seq.size()){
+            for(int i = seq.size()-1; i != n-1; ++i){
+                seq.push_back(seq[i]*r);
+                r *= rate2[0];
+            }
+        }
+        ans = seq[n-1];
+    }*/
+
+    for(int i = 0; i<seq.size(); ++i){
+        switch(i+1){
+        case 1:
+            cout << i+1 << "st" << "\t" << seq[i] << endl;
+            break;
+        case 2:
+            cout << i+1 << "nd" << "\t" << seq[i] << endl;
+            break;
+        case 3:
+            cout << i+1 << "rd" << "\t" << seq[i] << endl;
+            break;
+        default:
+            cout << i+1 << "th" << "\t" << seq[i] << endl;
+            break;
+        }
+    }
+    cout << endl;
+
+    return ans;
+}
+
+double Geometric::sum_of_geo(const vector<double>& seq, const int& n, double& nth_num)
+{
+    double a = seq[0];
+    double r = seq[1]/seq[0];
+    double sum = a*((1-pow(r,n))/(1-r));  // Sn = a(1-r^n/1-r)
+
+    return sum;
+}
